+++ fabric/src/main/java/com/gitlab/srcmc/rctmod/fabric/FabricCommon.java
@@ -20,10 +20,13 @@ package com.gitlab.srcmc.rctmod.fabric;
 import com.gitlab.srcmc.rctmod.ModCommon;
 import com.gitlab.srcmc.rctmod.ModRegistries;
 import com.gitlab.srcmc.rctmod.api.RCTMod;
+import com.gitlab.srcmc.rctmod.commands.arguments.TokenArgumentType;
 
 import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeConfigRegistry;
 import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeModConfigEvents;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.command.v2.ArgumentTypeRegistry;
+import net.minecraft.commands.synchronization.SingletonArgumentInfo;
 import net.neoforged.fml.config.ModConfig;
 
 public class FabricCommon implements ModInitializer {
@@ -37,6 +40,11 @@ public class FabricCommon implements ModInitializer {
         NeoForgeConfigRegistry.INSTANCE.register(ModCommon.MOD_ID, ModConfig.Type.SERVER, RCTMod.getInstance().getServerConfig().getSpec());
         NeoForgeModConfigEvents.loading(ModCommon.MOD_ID).register(FabricCommon::onConfigLoadingOrReloading);
         NeoForgeModConfigEvents.reloading(ModCommon.MOD_ID).register(FabricCommon::onConfigLoadingOrReloading);
+
+        ArgumentTypeRegistry.registerArgumentType(
+            ModRegistries.location("token_argument_type"),
+            TokenArgumentType.class,
+            SingletonArgumentInfo.contextFree(TokenArgumentType::token));
     }
 
     static void onConfigLoadingOrReloading(ModConfig config) {